*unused variables WILL be used in the next update!
transferFiles
This now writes the success indicator line to the log file only if BOTH files are downloaded successfully.

reconcileDownload
This now exists. Can see when it was last successful.
Future parts will determine the number and names of missing files and reinstantiate downloads accordingly.

writeConfigUpdate
This now exists. This writes the config file so that the next time the process runs it has that number that it references internally to see if it needs to download from the server (make the comparison so that it can determine if it's been working properly)

getLastSuccess
the iteration portion of the reconcileDownload function. Iterate through log files to search through failed and successful download attempts to figure out which was the last successful.